pyproject.toml
[tool.poetry]
name = "trading-network"
version = "0.1.0"
description = ""
authors = ["VladimirVlasov1982 <vladimir.vlasov82@mail.ru>"]
readme = "README.md"
packages = [{include = "trading_network"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
psycopg2 = "^2.9.5"
envparse = "^0.2.0"
djangorestframework = "^3.14.0"
django-filter = "^22.1"
drf-spectacular = "^0.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
